Move`usePageTitleSetter` hook to `hooks` directory
